requirements.txt
asgiref==3.7.2
Django==4.2.6
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
flake8==6.1.0
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
pytz==2023.3.post1
anyio==4.0.0
asgiref==3.7.2
certifi==2023.7.22
Django==4.2.6
djangorestframework==3.14.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
python-telegram-bot==20.6
pytz==2023.3.post1
sniffio==1.3.0
sqlparse==0.4.4
tzdata==2023.3
